# src/openapi.json
{"openapi":"3.1.0","info":{"title":"Bible App API","version":"0.0.0"},"paths":{"/health":{"get":{"tags":["health"],"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Health Health Get"}}}}}}},"/api/translations":{"get":{"tags":["api","translations"],"summary":"Get Translations","operationId":"get_translations_api_translations_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/TranslationInfo"},"type":"array","title":"Response Get Translations Api Translations Get"}}}}}}},"/api/translations/{translation_id}/metadata":{"get":{"tags":["api","translations"],"summary":"Get Translation Metadata","description":"Get complete metadata for a translation including:\n- All books with their IDs, names, and display names\n- Chapter counts per book\n- Verse counts per chapter\n- Total statistics","operationId":"get_translation_metadata_api_translations__translation_id__metadata_get","parameters":[{"name":"translation_id","in":"path","required":true,"schema":{"type":"string","title":"Translation Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TranslationMetadata"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/verses":{"get":{"tags":["api","verses"],"summary":"Get Verses","description":"Unified endpoint for Bible verse retrieval with flexible filtering.\n\n**Features:**\n- Text search with `q` parameter (exact or normalized matching)\n- Point retrieval using `book`, `chapter`, `verse` parameters\n- Range retrieval using `from_*` and `to_*` parameters\n- Combine search with location constraints\n- Support for multiple translations\n- Optional match highlighting\n\n**Parameter Types (mutually exclusive):**\n- Point queries: Use `book` [+ `chapter`] [+ `verse`]\n- Range queries: Use `from_book` + `to_*` parameters\n\n**Examples:**\n- Search: `/api/verses?translation_ids=eng-kjv&q=love`\n- Single verse: `/api/verses?translation_ids=eng-kjv&book=john&chapter=3&verse=16`\n- Full chapter: `/api/verses?translation_ids=eng-kjv&book=john&chapter=3`\n- Full book: `/api/verses?translation_ids=eng-kjv&book=john`\n- Verse range: `/api/verses?translation_ids=eng-kjv&from_book=john&from_chapter=3&from_verse=16&to_verse=18`\n- Chapter range: `/api/verses?translation_ids=eng-kjv&from_book=john&from_chapter=3&to_chapter=5`\n- Search in chapter: `/api/verses?translation_ids=eng-kjv&book=john&chapter=3&q=love`\n- Multiple translations: `/api/verses?translation_ids=eng-kjv,spa-rvr&book=john&chapter=3&verse=16`","operationId":"get_verses_api_verses_get","parameters":[{"name":"translation_ids","in":"query","required":true,"schema":{"type":"string","description":"Comma-separated translation IDs","title":"Translation Ids"},"description":"Comma-separated translation IDs"},{"name":"q","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Search query (optional, can be combined with location constraints)","title":"Q"},"description":"Search query (optional, can be combined with location constraints)"},{"name":"exact","in":"query","required":false,"schema":{"type":"boolean","description":"Use exact match searching","default":false,"title":"Exact"},"description":"Use exact match searching"},{"name":"highlight","in":"query","required":false,"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"description":"Highlight matches (default: true when q is provided)","title":"Highlight"},"description":"Highlight matches (default: true when q is provided)"},{"name":"book","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Book name (for point queries)","title":"Book"},"description":"Book name (for point queries)"},{"name":"chapter","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"description":"Chapter number (for point queries)","title":"Chapter"},"description":"Chapter number (for point queries)"},{"name":"verse","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"description":"Verse number (for point queries)","title":"Verse"},"description":"Verse number (for point queries)"},{"name":"from_book","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Starting book name (for range queries)","title":"From Book"},"description":"Starting book name (for range queries)"},{"name":"from_chapter","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"description":"Starting chapter (for range queries)","title":"From Chapter"},"description":"Starting chapter (for range queries)"},{"name":"from_verse","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"description":"Starting verse (for range queries)","title":"From Verse"},"description":"Starting verse (for range queries)"},{"name":"to_book","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Ending book name (for range queries)","title":"To Book"},"description":"Ending book name (for range queries)"},{"name":"to_chapter","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"description":"Ending chapter (for range queries)","title":"To Chapter"},"description":"Ending chapter (for range queries)"},{"name":"to_verse","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"description":"Ending verse (for range queries)","title":"To Verse"},"description":"Ending verse (for range queries)"},{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"description":"Page number","default":1,"title":"Page"},"description":"Page number"},{"name":"page_size","in":"query","required":false,"schema":{"type":"integer","maximum":1000,"minimum":1,"description":"Number of verses per page","default":30,"title":"Page Size"},"description":"Number of verses per page"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnifiedResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"BookInfo":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"displan_name":{"type":"string","title":"Displan Name"}},"type":"object","required":["id","name","displan_name"],"title":"BookInfo","description":"Book identification across all endpoints."},"BookMetadata":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"displan_name":{"type":"string","title":"Displan Name"},"chapter_count":{"type":"integer","title":"Chapter Count"},"chapters":{"items":{"$ref":"#/components/schemas/ChapterInfo"},"type":"array","title":"Chapters"}},"type":"object","required":["id","name","displan_name","chapter_count","chapters"],"title":"BookMetadata","description":"Book metadata extends BookInfo with structure details."},"ChapterInfo":{"properties":{"chapter":{"type":"integer","title":"Chapter"},"verse_count":{"type":"integer","title":"Verse Count"}},"type":"object","required":["chapter","verse_count"],"title":"ChapterInfo"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"PaginationInfo":{"properties":{"page":{"type":"integer","title":"Page"},"page_size":{"type":"integer","title":"Page Size"},"total_items":{"type":"integer","title":"Total Items"},"total_pages":{"type":"integer","title":"Total Pages"},"previous":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Previous"},"next":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Next"}},"type":"object","required":["page","page_size","total_items","total_pages","previous","next"],"title":"PaginationInfo","description":"Pagination metadata."},"TranslationContent":{"properties":{"translation_id":{"type":"string","title":"Translation Id"},"verses":{"items":{"$ref":"#/components/schemas/VerseData"},"type":"array","title":"Verses"}},"type":"object","required":["translation_id","verses"],"title":"TranslationContent"},"TranslationInfo":{"properties":{"public_id":{"type":"string","title":"Public Id"},"abbreviation":{"type":"string","title":"Abbreviation"},"full_name":{"type":"string","title":"Full Name"},"language_code":{"type":"string","title":"Language Code"}},"type":"object","required":["public_id","abbreviation","full_name","language_code"],"title":"TranslationInfo","description":"Translation identification and metadata."},"TranslationMetadata":{"properties":{"public_id":{"type":"string","title":"Public Id"},"abbreviation":{"type":"string","title":"Abbreviation"},"full_name":{"type":"string","title":"Full Name"},"language_code":{"type":"string","title":"Language Code"},"books":{"items":{"$ref":"#/components/schemas/BookMetadata"},"type":"array","title":"Books"},"total_books":{"type":"integer","title":"Total Books"},"total_chapters":{"type":"integer","title":"Total Chapters"},"total_verses":{"type":"integer","title":"Total Verses"}},"type":"object","required":["public_id","abbreviation","full_name","language_code","books","total_books","total_chapters","total_verses"],"title":"TranslationMetadata","description":"Translation metadata with complete structure."},"UnifiedResponse":{"properties":{"pagination":{"$ref":"#/components/schemas/PaginationInfo"},"results":{"items":{"$ref":"#/components/schemas/TranslationContent"},"type":"array","title":"Results"}},"type":"object","required":["pagination","results"],"title":"UnifiedResponse","description":"Unified response for verse retrieval with flexible filtering."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"VerseData":{"properties":{"book":{"$ref":"#/components/schemas/BookInfo"},"chapter":{"type":"integer","title":"Chapter"},"verse":{"type":"integer","title":"Verse"},"text":{"type":"string","title":"Text"}},"type":"object","required":["book","chapter","verse","text"],"title":"VerseData","description":"A Bible verse with complete reference and text."}}}}